Assume all songs are ranked on GS
Even when we have local scores for them.
This should resolve issues with SRPG leaderboards not loading for
"unranked" songs because we will make a request to GS every time.

As a bonus, all songs that will become ranked in the future but are
unranked now (and we have scores for them) will automatically start
displaying upstream scores.