Optimize songs list page
It's achieved by further denormalization of the database -- number of scores and unique players
is stored in a Song instance and updated on a score submission. Song was already modified when
there was a new highscore -- now it's always going to be updated when a new score is added for it.

Before it was a linear (to the number of scores and players in the database) operation to display
any page of the songs list. Now it should benefit from the database index when making a select.